create a updateVideoTime spec file
