+++ src/content/__tests__/content.spec.ts
@@ -1,7 +1,5 @@
 import { chrome } from 'jest-chrome';
 import { run, loadOptions } from '../content';
-import { updateVideoTime } from '../handle-video-player';
-import { ArrowKey } from '../types';
 import {
   DEFAULT_OPTIONS_MOCK,
   HTML_PLAYER_FULL,
@@ -24,37 +22,6 @@ describe('full run', () => {
       document.querySelectorAll('button.ml-custom-rewind-forward-buttons')
         ?.length
     ).toEqual(0);
-  });
-});
-
-describe('updateVideoTime', () => {
-  const videoElement = document.createElement('video');
-  it('should reduce to the currentTime video when ARROW_LEFT_KEY', () => {
-    videoElement.currentTime = 100;
-    updateVideoTime({
-      seconds: 30,
-      video: videoElement,
-      updateType: ArrowKey.ARROW_LEFT_KEY,
-    });
-    expect(videoElement.currentTime).toBe(70);
-  });
-  it('should add to the currentTime video when ARROW_RIGHT_KEY', () => {
-    videoElement.currentTime = 100;
-    updateVideoTime({
-      seconds: 30,
-      video: videoElement,
-      updateType: ArrowKey.ARROW_RIGHT_KEY,
-    });
-    expect(videoElement.currentTime).toBe(130);
-  });
-  it('should do nothing to the currentTime video when no arrow', () => {
-    videoElement.currentTime = 100;
-    updateVideoTime({
-      seconds: 30,
-      video: videoElement,
-      updateType: 'test_key' as any,
-    });
-    expect(videoElement.currentTime).toBe(100);
   });
 });
